+++ flask_jsonapi/resource_repositories.py
@@ -1,12 +1,13 @@
 from flask_jsonapi import descriptors
 from flask_jsonapi import exceptions
 from flask_jsonapi import filters_schema
+from flask_jsonapi import nested_repository
 
 from flask_jsonapi import resources
 
 
 class BaseResourceRepository:
-    def create(self, **kwargs):
+    def create(self, data, **kwargs):
         raise exceptions.NotImplementedMethod('Creating is not implemented.')
 
     def get_list(self, filters=None):
@@ -22,15 +23,15 @@ class BaseResourceRepository:
         raise exceptions.NotImplementedMethod('Updating is not implemented')
 
 
-
 class ResourceRepositoryViewSet:
     repository = BaseResourceRepository()
     schema = descriptors.NotImplementedProperty('schema')
     filter_schema = filters_schema.FilterSchema({})
     view_decorators = ()
     view_kwargs = None
+    nested = False
 
-    def __init__(self, *, repository=None, schema=None, filter_schema=None, view_decorators=None, view_kwargs=None):
+    def __init__(self, *, repository=None, schema=None, filter_schema=None, view_decorators=None, view_kwargs=None, nested=False):
         if repository:
             self.repository = repository
         if schema:
@@ -41,6 +42,10 @@ class ResourceRepositoryViewSet:
             self.view_decorators = view_decorators
         if view_kwargs:
             self.view_kwargs = view_kwargs
+        if nested:
+            self.nested = nested
+        if self.nested:
+            self.repository = self.extend_repository()
 
     def as_detail_view(self, view_name):
         return self.decorate(
@@ -48,8 +53,9 @@ class ResourceRepositoryViewSet:
         )
 
     def as_list_view(self, view_name):
+        cls = NestedResourceRepositoryListView if self.nested else ResourceRepositoryListView
         return self.decorate(
-            ResourceRepositoryListView.as_view(view_name, filter_schema=self.filter_schema, **self.get_views_kwargs())
+            cls.as_view(view_name, filter_schema=self.filter_schema, **self.get_views_kwargs())
         )
 
     def decorate(self, view):
@@ -61,8 +67,12 @@ class ResourceRepositoryViewSet:
         return {
             'schema': self.schema,
             'repository': self.repository,
+            'nested': self.nested,
             **(self.view_kwargs or {})
         }
+
+    def extend_repository(self):
+        return nested_repository.NestedRepository(repository=self.repository)
 
 
 class ResourceRepositoryViewMixin:
@@ -90,5 +100,13 @@ class ResourceRepositoryListView(ResourceRepositoryViewMixin, resources.Resource
     def read_many(self, filters):
         return self.repository.get_list(filters)
 
-    def create(self, data):
-        return self.repository.create(**data)
+    def create(self, data, **kwargs):
+        return self.repository.create(data, **kwargs)
+
+
+class NestedResourceRepositoryListView(ResourceRepositoryViewMixin, resources.NestedResourceList):
+    def read_many(self, filters):
+        return self.repository.get_list(filters)
+
+    def create(self, data, **kwargs):
+        return self.repository.create(data, **kwargs)